requirements.txt
python-docx
tqdm
PyMuPDF
scikit-learn
opencv-python
pandas
joblib
reportlab
numpy